backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching
redis==5.0.1
hiredis==2.3.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# AI/ML
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.37.2
timm==0.9.12
huggingface-hub==0.20.3
google-generativeai==0.3.2
pillow==10.2.0
opencv-python==4.9.0.80
librosa==0.10.1
soundfile==0.12.1
albumentations==1.3.1

# Google Cloud
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.17.0
google-cloud-logging==3.9.0
google-cloud-monitoring==2.18.0

# API Clients
httpx==0.26.0
aiohttp==3.9.3
tavily-python==0.3.3

# Validation & Serialization
python-magic==0.4.27
email-validator==2.1.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Rate Limiting & Middleware
slowapi==0.1.9
python-jose==3.3.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Utils
python-dotenv==1.0.0
tenacity==8.2.3
click==8.1.7
pyyaml==6.0.1
